feat : BadgeStackView를 대신할 BadgesView 생성 #3
